+++ wargaming/wows/ships_stats.go
@@ -11,6 +11,7 @@ type ShipsStatsOptions struct {
 	AccessToken *string `json:"access_token,omitempty"`
 	// Extra fields that will be added to the response. Valid values:
 	//
+	// "clan"
 	// "club"
 	// "oper_div"
 	// "oper_div_hard"
@@ -59,6 +60,107 @@ type ShipsStats struct {
 	AccountId *int `json:"account_id,omitempty"`
 	// Battles fought
 	Battles *int `json:"battles,omitempty"`
+	// Clan battles statistics.
+	// An extra field.
+	Clan *struct {
+		// Statistics of aircraft used
+		Aircraft *struct {
+			// Warships destroyed
+			Frags *int `json:"frags,omitempty"`
+			// Maximum number of enemy warships destroyed per battle
+			MaxFragsBattle *int `json:"max_frags_battle,omitempty"`
+		} `json:"aircraft,omitempty"`
+		// Potential damage caused by shells
+		ArtAgro *int `json:"art_agro,omitempty"`
+		// Battles fought
+		Battles *int `json:"battles,omitempty"`
+		// Base capture points
+		CapturePoints *int `json:"capture_points,omitempty"`
+		// Damage caused
+		DamageDealt *int `json:"damage_dealt,omitempty"`
+		// Damage caused by allies
+		DamageScouting *int `json:"damage_scouting,omitempty"`
+		// Draws
+		Draws *int `json:"draws,omitempty"`
+		// Base defense points
+		DroppedCapturePoints *int `json:"dropped_capture_points,omitempty"`
+		// Warships destroyed
+		Frags *int `json:"frags,omitempty"`
+		// Defeats
+		Losses *int `json:"losses,omitempty"`
+		// Main battery firing statistics
+		MainBattery *struct {
+			// Warships destroyed
+			Frags *int `json:"frags,omitempty"`
+			// Hits
+			Hits *int `json:"hits,omitempty"`
+			// Maximum number of enemy warships destroyed per battle
+			MaxFragsBattle *int `json:"max_frags_battle,omitempty"`
+			// Shots fired
+			Shots *int `json:"shots,omitempty"`
+		} `json:"main_battery,omitempty"`
+		// Maximum Damage caused per battle
+		MaxDamageDealt *int `json:"max_damage_dealt,omitempty"`
+		// Most damage caused by allies to enemy ships spotted by the player
+		MaxDamageScouting *int `json:"max_damage_scouting,omitempty"`
+		// Maximum number of enemy warships destroyed per battle
+		MaxFragsBattle *int `json:"max_frags_battle,omitempty"`
+		// Maximum number of aircraft destroyed per battle
+		MaxPlanesKilled *int `json:"max_planes_killed,omitempty"`
+		// Most ships detected
+		MaxShipsSpotted *int `json:"max_ships_spotted,omitempty"`
+		// Most potential damage caused by ammunition that hit or fell near the ship
+		MaxTotalAgro *int `json:"max_total_agro,omitempty"`
+		// Maximum Experience Points per battle, including XP earned with Premium Account
+		MaxXp *int `json:"max_xp,omitempty"`
+		// Enemy aircraft destroyed
+		PlanesKilled *int `json:"planes_killed,omitempty"`
+		// Statistics of ramming enemy warships
+		Ramming *struct {
+			// Warships destroyed
+			Frags *int `json:"frags,omitempty"`
+			// Maximum number of enemy warships destroyed per battle
+			MaxFragsBattle *int `json:"max_frags_battle,omitempty"`
+		} `json:"ramming,omitempty"`
+		// Secondary armament firing statistics
+		SecondBattery *struct {
+			// Warships destroyed
+			Frags *int `json:"frags,omitempty"`
+			// Hits
+			Hits *int `json:"hits,omitempty"`
+			// Maximum number of enemy warships destroyed per battle
+			MaxFragsBattle *int `json:"max_frags_battle,omitempty"`
+			// Shots fired
+			Shots *int `json:"shots,omitempty"`
+		} `json:"second_battery,omitempty"`
+		// Ships spotted by the player first
+		ShipsSpotted *int `json:"ships_spotted,omitempty"`
+		// Battles survived
+		SurvivedBattles *int `json:"survived_battles,omitempty"`
+		// Victories in battles survived
+		SurvivedWins *int `json:"survived_wins,omitempty"`
+		// Total number of base capture points earned by the player's team
+		TeamCapturePoints *int `json:"team_capture_points,omitempty"`
+		// Total number of base defense points earned by the player's team
+		TeamDroppedCapturePoints *int `json:"team_dropped_capture_points,omitempty"`
+		// Potential damage caused by torpedoes
+		TorpedoAgro *int `json:"torpedo_agro,omitempty"`
+		// Statistics of attacking targets with torpedoes
+		Torpedoes *struct {
+			// Warships destroyed
+			Frags *int `json:"frags,omitempty"`
+			// Hits
+			Hits *int `json:"hits,omitempty"`
+			// Maximum number of enemy warships destroyed per battle
+			MaxFragsBattle *int `json:"max_frags_battle,omitempty"`
+			// Shots fired
+			Shots *int `json:"shots,omitempty"`
+		} `json:"torpedoes,omitempty"`
+		// Victories
+		Wins *int `json:"wins,omitempty"`
+		// Total Experience Points , including XP earned with Premium Account
+		Xp *int `json:"xp,omitempty"`
+	} `json:"clan,omitempty"`
 	// Statistics in Team battles.
 	// An extra field.
 	Club *struct {